Refactor persistence layer and update database schema
Replaced `StoreContext` with `StoreDbContext` and updated its usage across the project, including dependency injection, repositories, and unit of work. Introduced `StoreDbInitializer` to replace `StoreContextInitializer` for database initialization and seeding.

Reintroduced entity configurations for `Product`, `ProductBrand`, and `ProductCategory` with updated property constraints and relationships. Added a new `NormalizedName` column to the `Products` table with a maximum length constraint.

Updated migrations and `StoreContextModelSnapshot` to reflect schema changes. Reintroduced JSON seed data files for `ProductBrand`, `ProductCategory`, and `Product` entities.

Updated `Talabat.Infrastructure.Persistence.csproj` to include necessary EF Core package references. Ensured the persistence layer is consistent, maintainable, and aligned with application requirements.
