+++ Talabat.Infrastructure.Persistence/AssemblyInformation.cs
@@ -1,4 +1,4 @@
-﻿namespace Talabat.Infrastructure.Persistence.Data
+﻿namespace Talabat.Infrastructure.Persistence
 {
     public static class AssemblyInformation
     {
