+++ Talabat.Infrastructure.Persistence/DependencyInjection.cs
@@ -2,17 +2,18 @@
 using Microsoft.Extensions.DependencyInjection;
 using Talabat.Domain.Contracts.Infrastructure;
 using Talabat.Domain.Contracts.Presistence;
+using Talabat.Infrastructure.Persistence._Data;
 namespace Talabat.Infrastructure.Persistence
 {
     public static class DependencyInjection
     {
         public static IServiceCollection AddPersistenceServices(this IServiceCollection services, IConfiguration configuration)
         {
-            services.AddDbContext<StoreContext>(options =>
+            services.AddDbContext<StoreDbContext>(options =>
                          options.UseLazyLoadingProxies()
                          .UseSqlServer(configuration.GetConnectionString("StoreConnection")));
 
-            services.AddScoped<IStoreContextIntiializer, StoreContextInitializer>();
+            services.AddScoped<IStoreContextIntiializer, StoreDbInitializer>();
             services.AddScoped(typeof(IUnitOfWork), typeof(UnitOfWork.UnitOfWork));
           
             return services;
